+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Unstyled to styled state transition when the class has transition property.
+- Responsive style back to style without breakpoint when the page reloaded on the different screen.
+
 ## [1.0.2] - [2024-7-27]
 
 ### Fixed